requirements.txt
streamlit==1.36.0
sentence-transformers==3.0.1
chromadb==0.4.24
groq==0.13.0
python-dotenv==1.0.1
transformers==4.40.0
torch==2.3.0
nltk==3.8.1
rank-bm25==0.2.2
scikit-learn==1.3.0
pyyaml==6.0.1
networkx==3.1
httpx==0.27.2
pipdeptree==2.23.4
numpy==1.26.4
onnxruntime==1.21.0